# openapi.yaml
openapi: 3.1.0
info:
  title: PFF Fantasy League API
  version: "1.0.0"
  description: |
    REST interface for the PFF Fantasy League toolkit. All requests must include
    the header `X-API-Key` matching the `PFF_API_KEY` environment variable.
servers:
  - url: https://pff-fantasy-analyzer.onrender.com
    description: Production
  - url: http://localhost:8000
    description: Local development
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    ConfigKnobs:
      type: object
      description: Scoring knobs currently in effect. Values are floats between 0 and 1 unless noted.
      additionalProperties: false
      properties:
        fuzzy_cutoff:
          type: number
          format: float
          description: Minimum similarity required when matching roster names.
        bench_ovar_beta:
          type: number
          format: float
          description: Weight applied to overall rank advantage when scoring bench players.
        combined_starters_weight:
          type: number
          format: float
          description: Share of the combined score attributed to starter strength.
        combined_bench_weight:
          type: number
          format: float
          description: Share of the combined score attributed to bench depth.
        projection_scale_beta:
          type: number
          format: float
          description: Influence of projection z-scores when rescaling rankings.
        bench_z_fallback_threshold:
          type: number
          format: float
          description: Threshold that swaps to percentile-based bench scoring when variance is low.
        bench_percentile_clamp:
          type: number
          format: float
          description: Clamp applied to percentile-based z-scores to limit outlier influence.
      required:
        - fuzzy_cutoff
        - bench_ovar_beta
        - combined_starters_weight
        - combined_bench_weight
        - projection_scale_beta
        - bench_z_fallback_threshold
        - bench_percentile_clamp
...
    LeaderboardEntry:
      type: object
      properties:
        team: {type: string}
        value: {type: number}
    StarterDetail:
      type: object
      properties:
        name: {type: string}
        position: {type: string, nullable: true}
        csvName: {type: string, nullable: true}
        rank: {type: integer, nullable: true}
        posRank: {type: integer, nullable: true}
        projection: {type: number, nullable: true}
        vor: {type: number, nullable: true}
    BenchDetail:
      type: object
      properties:
        name: {type: string}
        position: {type: string, nullable: true}
        rank: {type: integer, nullable: true}
        posRank: {type: integer, nullable: true}
        projection: {type: number, nullable: true}
        vor: {type: number, nullable: true}
        benchScore: {type: number, nullable: true}
        oVAR: {type: number, nullable: true}
    TeamDetail:
      type: object
      properties:
        team: {type: string}
        starters:
          type: array
          items: {$ref: '#/components/schemas/StarterDetail'}
        bench:
          type: array
          items: {$ref: '#/components/schemas/BenchDetail'}
        benchLimit: {type: integer, nullable: true}
    TeamEvaluation:
      type: object
      properties:
        team: {type: string}
        combinedScore: {type: number}
        starterVOR: {type: number}
        benchScore: {type: number}
        starterProjection: {type: number}
    EvaluateRequest:
      type: object
      properties:
        projectionScaleBeta: {type: number, nullable: true}
        replacementSkipPct: {type: number, nullable: true}
        replacementWindow: {type: integer, nullable: true}
        benchOvarBeta: {type: number, nullable: true}
        combinedStartersWeight: {type: number, nullable: true}
        combinedBenchWeight: {type: number, nullable: true}
        benchZFallbackThreshold: {type: number, nullable: true}
        benchPercentileClamp: {type: number, nullable: true}
        scarcitySampleStep: {type: number, nullable: true}
        rankingsPath: {type: string, nullable: true}
        projectionsPath: {type: string, nullable: true}
        supplementalPath: {type: string, nullable: true}
        rosters:
          type: object
          additionalProperties:
            type: object
            additionalProperties:
              type: array
              items: {type: string}
        includeDetails: {type: boolean, default: false}
        benchLimit: {type: integer, nullable: true}
    EvaluateResponse:
      type: object
      properties:
        evaluatedAt: {type: string, format: date-time}
        playerCount: {type: integer}
        teams:
          type: array
          items: {$ref: '#/components/schemas/TeamEvaluation'}
        leaderboards:
          type: object
          additionalProperties:
            type: array
            items: {$ref: '#/components/schemas/LeaderboardEntry'}
        settings:
          $ref: '#/components/schemas/ConfigKnobs'
        replacementPoints:
          type: object
          additionalProperties: {type: number}
        replacementTargets:
          type: object
          additionalProperties: {type: number}
        scarcitySamples:
          type: object
          additionalProperties:
            type: array
            items:
              type: object
              properties:
                slot: {type: number}
                projection: {type: number}
        zeroSum:
          $ref: '#/components/schemas/ZeroSumResponse'
        details:
          type: array
          items: {$ref: '#/components/schemas/TeamDetail'}
          nullable: true
    ZeroSumEntry:
      type: object
      properties:
        team: {type: string}
        value: {type: number}
        share: {type: number}
        surplus: {type: number}
      required: [team, value, share, surplus]
    ZeroSumGroup:
      type: object
      properties:
        total: {type: number}
        baseline: {type: number}
        shareSum: {type: number}
        surplusSum: {type: number}
        entries:
          type: array
          items: {$ref: '#/components/schemas/ZeroSumEntry'}
      required: [total, baseline, shareSum, surplusSum, entries]
    ZeroSumCombinedGroup:
      allOf:
        - $ref: '#/components/schemas/ZeroSumGroup'
        - type: object
          properties:
            weights:
              type: object
              additionalProperties: {type: number}
          required: [weights]
    ZeroSumScarcityMetric:
      type: object
      properties:
        deficit: {type: number}
        pressure: {type: number}
      required: [deficit, pressure]
    ZeroSumConcentrationRisk:
      type: object
      properties:
        starterPositions:
          type: object
          additionalProperties: {type: number}
        benchPositions:
          type: object
          additionalProperties: {type: number}
        slotShares:
          type: object
          additionalProperties: {type: number}
        flexShare: {type: number}
        herfindahl:
          type: object
          properties:
            starters: {type: number}
            bench: {type: number}
            slots: {type: number}
          additionalProperties: false
      required: [starterPositions, benchPositions, slotShares, flexShare, herfindahl]
    ZeroSumAnalyticsTeam:
      type: object
      properties:
        scarcityPressure:
          type: object
          additionalProperties: {$ref: '#/components/schemas/ZeroSumScarcityMetric'}
        concentrationRisk:
          $ref: '#/components/schemas/ZeroSumConcentrationRisk'
      required: [scarcityPressure, concentrationRisk]
    ZeroSumAnalytics:
      type: object
      properties:
        teams:
          type: object
          additionalProperties: {$ref: '#/components/schemas/ZeroSumAnalyticsTeam'}
        league:
          type: object
          properties:
            highPressurePositions:
              type: array
              items:
                type: object
                properties:
                  position: {type: string}
                  aggregateDeficit: {type: number}
                required: [position, aggregateDeficit]
          additionalProperties: false
      required: [teams, league]
    ZeroSumResponse:
      type: object
      properties:
        teamCount: {type: integer}
        starters: {$ref: '#/components/schemas/ZeroSumGroup'}
        bench: {$ref: '#/components/schemas/ZeroSumGroup'}
        combined: {$ref: '#/components/schemas/ZeroSumCombinedGroup'}
        positions:
          type: object
          additionalProperties: {$ref: '#/components/schemas/ZeroSumGroup'}
        benchPositions:
          type: object
          additionalProperties: {$ref: '#/components/schemas/ZeroSumGroup'}
        slots:
          type: object
          additionalProperties: {$ref: '#/components/schemas/ZeroSumGroup'}
        flex:
          $ref: '#/components/schemas/ZeroSumGroup'
          nullable: true
        analytics:
          $ref: '#/components/schemas/ZeroSumAnalytics'
      required: [teamCount, starters, bench, combined, positions, benchPositions, slots, analytics]
    ZeroSumShiftMetrics:
      type: object
      properties:
        shareDelta: {type: number}
        surplusDelta: {type: number}
        valueDelta: {type: number}
      required: [shareDelta, surplusDelta, valueDelta]
    ZeroSumShift:
      type: object
      properties:
        combined:
          type: object
          additionalProperties: {$ref: '#/components/schemas/ZeroSumShiftMetrics'}
        starters:
          type: object
          additionalProperties: {$ref: '#/components/schemas/ZeroSumShiftMetrics'}
        bench:
          type: object
          additionalProperties: {$ref: '#/components/schemas/ZeroSumShiftMetrics'}
        positions:
          type: object
          additionalProperties:
            type: object
            additionalProperties: {$ref: '#/components/schemas/ZeroSumShiftMetrics'}
        benchPositions:
          type: object
          additionalProperties:
            type: object
            additionalProperties: {$ref: '#/components/schemas/ZeroSumShiftMetrics'}
        slots:
          type: object
          additionalProperties:
            type: object
            additionalProperties: {$ref: '#/components/schemas/ZeroSumShiftMetrics'}
        flex:
          type: object
          additionalProperties: {$ref: '#/components/schemas/ZeroSumShiftMetrics'}
        summary: {type: object}
      required: [combined, starters, bench, positions, benchPositions, slots, flex, summary]
    EvaluateDeltaSnapshot:
      type: object
      properties:
        teams:
          type: array
          items: {$ref: '#/components/schemas/TeamEvaluation'}
        zeroSum:
          $ref: '#/components/schemas/ZeroSumResponse'
      required: [teams, zeroSum]
    EvaluateDeltaResponse:
      type: object
      properties:
        evaluatedAt: {type: string, format: date-time}
        baseline: {$ref: '#/components/schemas/EvaluateDeltaSnapshot'}
        scenario: {$ref: '#/components/schemas/EvaluateDeltaSnapshot'}
        zeroSumShift: {$ref: '#/components/schemas/ZeroSumShift'}
      required: [evaluatedAt, baseline, scenario, zeroSumShift]
...
    TradePiece:
      type: object
      properties:
        group: {type: string}
        name: {type: string}
      required: [group, name]
    TradeTeamResult:
      type: object
      properties:
        team: {type: string}
        baseline: {type: number}
        postTrade: {type: number}
        delta: {type: number}
    TradeEvaluateRequest:
      type: object
      properties:
        teamA: {type: string}
        teamB: {type: string}
        sendA:
          type: array
          items: {$ref: '#/components/schemas/TradePiece'}
        sendB:
          type: array
          items: {$ref: '#/components/schemas/TradePiece'}
        includeDetails: {type: boolean, default: true}
        benchLimit: {type: integer, nullable: true}
      required: [teamA, teamB]
    TradeEvaluateResponse:
      type: object
      properties:
        evaluatedAt: {type: string, format: date-time}
        teams:
          type: array
          items: {$ref: '#/components/schemas/TradeTeamResult'}
        combinedScores:
          type: object
          additionalProperties: {type: number}
        replacementPoints:
          type: object
          additionalProperties: {type: number}
        replacementTargets:
          type: object
          additionalProperties: {type: number}
        starterVOR:
          type: object
          additionalProperties: {type: number}
        benchTotals:
          type: object
          additionalProperties: {type: number}
        zeroSumBefore:
          $ref: '#/components/schemas/ZeroSumResponse'
        zeroSumAfter:
          $ref: '#/components/schemas/ZeroSumResponse'
        zeroSumShift:
          $ref: '#/components/schemas/ZeroSumShift'
        leaderboards:
          type: object
          additionalProperties:
            type: array
            items: {$ref: '#/components/schemas/LeaderboardEntry'}
        details:
          type: array
          items: {$ref: '#/components/schemas/TeamDetail'}
          nullable: true
...
paths:
...
  /evaluate:
    post:
      tags: [league]
      summary: Evaluate league
      operationId: evaluateLeague
      description: |
        Submit an `EvaluateRequest` JSON body to tweak knobs and hypothetical rosters. For convenience,
        `includeDetails` and `benchLimit` may also be supplied as query parameters when you only need
        to toggle those flags.
      x-openai-isConsequential: false
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EvaluateRequest'
            examples:
              default:
                summary: Baseline evaluation
                value: {}
              withDetails:
                summary: Include details with custom bench limit
                value:
                  includeDetails: true
                  benchLimit: 3
      responses:
        '200':
          description: Evaluation result
          content:
          application/json:
            schema:
              $ref: '#/components/schemas/EvaluateResponse'
  /evaluate/delta:
    post:
      tags: [league]
      summary: Compare baseline and modified rosters
      operationId: evaluateLeagueDelta
      description: |
        Run two evaluations (baseline vs. hypothetical roster override) and return both results alongside
        a zero-sum shift summary (share/surplus/value deltas with winners/losers callouts).
      security:
        - ApiKeyAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EvaluateRequest'
      responses:
        '200':
          description: Baseline and scenario comparison
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvaluateDeltaResponse'
...
  /trade/evaluate:
    post:
      tags: [trade]
      summary: Evaluate trade proposal
      operationId: evaluateTrade
      x-openai-isConsequential: false
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TradeEvaluateRequest'
            example:
              teamA: JohnBurrows_School
              teamB: KWood_Super_Marios
              sendA:
                - {group: WR, name: George Pickens}
              sendB:
                - {group: WR, name: Drake London}
              includeDetails: true
      responses:
        '200':
          description: Trade impact summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TradeEvaluateResponse'
...
security:
  - ApiKeyAuth: []